Use consistent s32/u32 (#2964)
* cleanup long usage

* fix regression

+++ include/JSystem/JGadget/define.h
@@ -15,14 +15,14 @@ public:
     JGadget_outMessage(MessageFunc fn, const char* file, int line);
     ~JGadget_outMessage();
 
-    JGadget_outMessage& operator<<(int param_1) { return *this << (signed long)param_1; }
+    JGadget_outMessage& operator<<(int param_1) { return *this << (s32)param_1; }
     JGadget_outMessage& operator<<(u16);
-    JGadget_outMessage& operator<<(unsigned int);
+    JGadget_outMessage& operator<<(uint);
     JGadget_outMessage& operator<<(u8 param_1) { return *this << (char)param_1; }
     JGadget_outMessage& operator<<(const char* str);
     JGadget_outMessage& operator<<(char);
-    JGadget_outMessage& operator<<(signed long);
-    JGadget_outMessage& operator<<(unsigned long);
+    JGadget_outMessage& operator<<(s32);
+    JGadget_outMessage& operator<<(u32);
     JGadget_outMessage& operator<<(const void*);
 
 private:
