+++ include/JSystem/JUtility/JUTException.h
@@ -152,7 +152,7 @@ STATIC_ASSERT(sizeof(JUTException) == 0xA4);
 */
 struct JUTWarn {
     JUTWarn& operator<<(const char*) { return *this; }
-    JUTWarn& operator<<(long) { return *this; }
+    JUTWarn& operator<<(s32) { return *this; }
 };
 
 #endif /* JUTEXCEPTION_H */
